backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
python-dotenv
google-generativeai
PyMuPDF
reportlab
firebase-admin